delete last n history entries

+++ cli/history.go
@@ -56,18 +56,43 @@ func list(service service.HistoryServiceIface) *v2.Command {
 			}
 
 			selected := fzfHistory(history)
+			// TODO: Implement copy to clipboard
 			fmt.Println(selected)
-
 			return nil
 		},
 	}
 }
 
 func delete(service service.HistoryServiceIface) *v2.Command {
+	lastFlag := &v2.IntFlag{
+		Name:  "last",
+		Usage: "purge the last (n) entries",
+	}
+
 	return &v2.Command{
 		Name:  "delete",
 		Usage: "delete an entry",
+		Flags: []v2.Flag{
+			lastFlag,
+		},
 		Action: func(ctx *v2.Context) error {
+			last := ctx.Int("last")
+
+			if last > 0 {
+				history, err := service.GetRecentHistory(last)
+
+				if err != nil {
+					return err
+				}
+
+				for _, entry := range history {
+					service.DeleteEntry(entry.ID)
+				}
+
+				fmt.Printf("Last (%d) have been successfully purged.\n", last)
+				return nil
+			}
+
 			history, err := service.GetAllHistory()
 
 			if err != nil {
@@ -75,6 +100,11 @@ func delete(service service.HistoryServiceIface) *v2.Command {
 			}
 
 			selected := fzfHistory(history)
+
+			if selected == nil {
+				fmt.Println("No entry selected.")
+				return nil
+			}
 
 			err = service.DeleteEntry(selected.ID)
 
@@ -89,7 +119,6 @@ func delete(service service.HistoryServiceIface) *v2.Command {
 }
 
 func fzfHistory(history []models.SearchHistoryEvent) *models.SearchHistoryEvent {
-
 	index, err := fzf.FindMulti(
 		history,
 		func(i int) string {
@@ -111,6 +140,9 @@ func fzfHistory(history []models.SearchHistoryEvent) *models.SearchHistoryEvent 
 		return nil
 	}
 
-	return &history[index[0]]
+	if len(index) > 0 {
+		return &history[index[0]]
+	}
 
+	return nil
 }
